Move deploy package out of properties package
