+++ java/clients/src/main/java/sleeper/clients/deploy/DeployNewInstance.java
@@ -35,9 +35,9 @@ import sleeper.clients.util.EcrRepositoryCreator;
 import sleeper.clients.util.cdk.CdkCommand;
 import sleeper.clients.util.cdk.InvokeCdkForInstance;
 import sleeper.configuration.properties.S3InstanceProperties;
+import sleeper.core.deploy.DeployInstanceConfiguration;
+import sleeper.core.deploy.DeployInstanceConfigurationFromTemplates;
 import sleeper.core.properties.SleeperPropertiesValidationReporter;
-import sleeper.core.properties.deploy.DeployInstanceConfiguration;
-import sleeper.core.properties.deploy.DeployInstanceConfigurationFromTemplates;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.local.SaveLocalProperties;
 import sleeper.core.properties.table.TableProperties;
