+++ java/clients/src/main/java/sleeper/clients/deploy/PopulateInstanceProperties.java
@@ -20,8 +20,8 @@ import com.amazonaws.services.securitytoken.AWSSecurityTokenService;
 import com.amazonaws.services.securitytoken.model.GetCallerIdentityRequest;
 import software.amazon.awssdk.regions.providers.AwsRegionProvider;
 
-import sleeper.core.properties.deploy.DeployInstanceConfiguration;
-import sleeper.core.properties.deploy.SleeperScheduleRule;
+import sleeper.core.deploy.DeployInstanceConfiguration;
+import sleeper.core.deploy.SleeperScheduleRule;
 import sleeper.core.properties.instance.InstanceProperties;
 
 import java.nio.file.Path;
