+++ java/core/src/main/java/sleeper/core/deploy/DeployInstanceConfigurationFromTemplates.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.core.properties.deploy;
+package sleeper.core.deploy;
 
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.local.LoadLocalProperties;
