+++ java/core/src/main/java/sleeper/core/deploy/GeneratePropertiesTemplates.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.properties.deploy;
+package sleeper.core.deploy;
 
 import sleeper.core.properties.PropertyGroup;
 import sleeper.core.properties.SleeperProperties;
