+++ java/core/src/test/java/sleeper/core/deploy/GeneratePropertiesTemplatesTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.properties.deploy;
+package sleeper.core.deploy;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
