+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/testutil/LocalStackTestInstance.java
@@ -15,7 +15,7 @@
  */
 package sleeper.systemtest.drivers.testutil;
 
-import sleeper.core.properties.deploy.DeployInstanceConfiguration;
+import sleeper.core.deploy.DeployInstanceConfiguration;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.schema.Field;
